Dynamically add/remove extensions' help
Move extensions' help management from ExtensionFactory to ExtensionHelp,
to dynamically add/remove the help when the (main) help add-on is
installed/uninstalled, not just when ZAP is started or the extension
removed, ensuring the help is always available (whenever possible).
