+++ src/org/zaproxy/zap/extension/help/ExtensionHelp.java
@@ -26,6 +26,8 @@ import java.awt.event.KeyEvent;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Map.Entry;
+import java.util.function.Function;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -46,6 +48,7 @@ import javax.swing.UIManager;
 
 import org.apache.log4j.Logger;
 import org.parosproxy.paros.Constant;
+import org.parosproxy.paros.extension.Extension;
 import org.parosproxy.paros.extension.ExtensionAdaptor;
 import org.parosproxy.paros.extension.ExtensionHook;
 import org.parosproxy.paros.extension.ViewDelegate;
@@ -129,7 +132,7 @@ public class ExtensionHelp extends ExtensionAdaptor {
 	 */
 	private static WeakHashMap<JComponent, String> componentsWithHelp;
 
-	private static Map<AddOn, HelpSet> addOnHelpSets = new HashMap<>();
+	private static Map<AddOn, List<HelpSet>> addOnHelpSets = new HashMap<>();
 
 	private static final Logger logger = Logger.getLogger(ExtensionHelp.class);
 	
@@ -245,6 +248,8 @@ public class ExtensionHelp extends ExtensionAdaptor {
 	}
 
 	private static void loadAddOnHelpSet(AddOn addOn) {
+		addOn.getLoadedExtensions().forEach(ExtensionHelp::loadExtensionHelpSet);
+
 		AddOn.HelpSetData helpSetData = addOn.getHelpSetData();
 		if (helpSetData.isEmpty()) {
 			return;
@@ -253,7 +258,7 @@ public class ExtensionHelp extends ExtensionAdaptor {
 		ClassLoader classLoader = addOn.getClassLoader();
 		URL helpSetUrl = LocaleUtils.findResource(
 				helpSetData.getBaseName(),
-				ExtensionHelp.HELP_SET_FILE_EXTENSION,
+				HELP_SET_FILE_EXTENSION,
 				helpSetData.getLocaleToken(),
 				Constant.getLocale(),
 				classLoader::getResource);
@@ -269,12 +274,51 @@ public class ExtensionHelp extends ExtensionAdaptor {
 
 		try {
 			logger.debug("Loading help for '" + addOn.getId() + "' add-on and merging with core help.");
-			HelpSet addOnHelpSet = new HelpSet(classLoader, helpSetUrl);
-			hb.getHelpSet().add(addOnHelpSet);
-			addOnHelpSets.put(addOn, addOnHelpSet);
+			addHelpSet(addOn, new HelpSet(classLoader, helpSetUrl));
 		} catch (HelpSetException e) {
 			logger.error("An error occured while adding help for '" + addOn.getId() + "' add-on:", e);
 		}
+	}
+
+	private static void addHelpSet(AddOn addOn, HelpSet helpSet) {
+		hb.getHelpSet().add(helpSet);
+		addOnHelpSets.computeIfAbsent(addOn, k -> new ArrayList<>()).add(helpSet);
+	}
+
+	private static void loadExtensionHelpSet(Extension ext) {
+		URL helpSetUrl = getExtensionHelpSetUrl(ext);
+		if (helpSetUrl != null) {
+			try {
+				logger.debug("Load help files for extension '" + ext.getName() + "' and merge with core help.");
+				addHelpSet(ext.getAddOn(), new HelpSet(ext.getClass().getClassLoader(), helpSetUrl));
+			} catch (HelpSetException e) {
+				logger.error(
+						"An error occured while adding help file of extension '" + ext.getName() + "': " + e.getMessage(),
+						e);
+			}
+		}
+	}
+
+	private static URL getExtensionHelpSetUrl(Extension extension) {
+		String extensionPackage = extension.getClass().getPackage().getName();
+		String localeToken = "%LC%";
+		Function<String, URL> getResource = extension.getClass().getClassLoader()::getResource;
+		URL helpSetUrl = LocaleUtils.findResource(
+				extensionPackage + ".resources.help" + localeToken + "." + HELP_SET_FILE_NAME,
+				HELP_SET_FILE_EXTENSION,
+				localeToken,
+				Constant.getLocale(),
+				getResource);
+		if (helpSetUrl == null) {
+			// Search in old location
+			helpSetUrl = LocaleUtils.findResource(
+					extensionPackage + ".resource.help" + localeToken + "." + HELP_SET_FILE_NAME,
+					HELP_SET_FILE_EXTENSION,
+					localeToken,
+					Constant.getLocale(),
+					getResource);
+		}
+		return helpSetUrl;
 	}
 
 	/**
@@ -466,8 +510,8 @@ public class ExtensionHelp extends ExtensionAdaptor {
             if (findHelpSetUrl() == null) {
                 setHelpEnabled(false);
             } else {
-                addOnHelpSets.computeIfPresent(addOn, (k, helpset) -> {
-                    EventQueue.invokeLater(() -> hb.getHelpSet().remove(helpset));
+                addOnHelpSets.computeIfPresent(addOn, (k, helpsets) -> {
+                    EventQueue.invokeLater(() -> helpsets.forEach(hb.getHelpSet()::remove));
                     return null;
                 });
             }
